Fix test case naming: p=cards/2 for single-node, p=P_cards+D_cards for PD separation
- GLM single-node: tp=16 -> 8p, tp=32 -> 16p
- GLM PD separation: P(tp=32)+D(tp=32) -> 1p1d_64p
- Kimi PD separation: P(tp=16)+D(tp=16) -> 1p1d_32p, 2P(tp=16)+D(tp=32) -> 2p1d_64p
- Update Kimi single-node perf test params to match issue 8 script
