+++ .github/workflows/full-test-npu.yml
@@ -450,17 +450,17 @@ jobs:
           #  runner: linux-aarch64-a3-16
           #  test_case: test/registered/ascend/performance/test_npu_qwen3_5_397b_w4a8_16p_in64k_out1k_prefix_50ms.py
           #  test_type: 'perf'
-          - name: test_npu_glm5_1_w4a8_16p_gpqa
+          - name: test_npu_glm5_1_w4a8_8p_gpqa
             runner: linux-aarch64-a3-16
-            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_16p_gpqa.py
+            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_8p_gpqa.py
             test_type: 'accuracy'
-          - name: test_npu_glm5_1_w4a8_16p_aime2025
+          - name: test_npu_glm5_1_w4a8_8p_aime2025
             runner: linux-aarch64-a3-16
-            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_16p_aime2025.py
+            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_8p_aime2025.py
             test_type: 'accuracy'
-          - name: test_npu_glm5_1_w4a8_16p_in3k5_out1k5_50ms
+          - name: test_npu_glm5_1_w4a8_8p_in3k5_out1k5_50ms
             runner: linux-aarch64-a3-16
-            test_case: test/registered/ascend/performance/test_npu_glm5_1_w4a8_16p_in3k5_out1k5_50ms.py
+            test_case: test/registered/ascend/performance/test_npu_glm5_1_w4a8_8p_in3k5_out1k5_50ms.py
             test_type: 'perf'
     uses: ./.github/workflows/nightly-test-npu-e2e-single-node.yml
     with:
@@ -488,90 +488,91 @@ jobs:
             test_case: test/registered/ascend/accuracy/test_ascend_qwen3_235b_w8a8_1p1d_24p_accuracy.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'separation'
-          - name: test_npu_glm5_1_w4a8_32p_gpqa
+          - name: test_npu_glm5_1_w4a8_16p_gpqa
             prefill_size: 2
             decode_size: 0
             router_size: 0
-            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_32p_gpqa.py
+            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_16p_gpqa.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'mix'
-          - name: test_npu_glm5_1_w4a8_32p_aime2025
+          - name: test_npu_glm5_1_w4a8_16p_aime2025
             prefill_size: 2
             decode_size: 0
             router_size: 0
-            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_32p_aime2025.py
+            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_16p_aime2025.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'mix'
-          - name: test_npu_glm5_1_w4a8_32p_in3k5_out1k5_50ms
+          - name: test_npu_glm5_1_w4a8_16p_in3k5_out1k5_50ms
             prefill_size: 2
             decode_size: 0
             router_size: 0
-            test_case: test/registered/ascend/performance/test_npu_glm5_1_w4a8_32p_in3k5_out1k5_50ms.py
+            test_case: test/registered/ascend/performance/test_npu_glm5_1_w4a8_16p_in3k5_out1k5_50ms.py
             test_type: 'perf'
             prefill_decode_deployment: 'mix'
-          - name: test_npu_glm5_1_w4a8_1p1d_32p_gpqa
+          # GLM 1p1d_64p (P: tp=32,nnodes=2 + D: tp=32,nnodes=2, total 64 cards)
+          - name: test_npu_glm5_1_w4a8_1p1d_64p_gpqa
             prefill_size: 2
             decode_size: 2
             router_size: 1
-            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_1p1d_32p_gpqa.py
+            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_1p1d_64p_gpqa.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'separation'
-          - name: test_npu_glm5_1_w4a8_1p1d_32p_aime2025
+          - name: test_npu_glm5_1_w4a8_1p1d_64p_aime2025
             prefill_size: 2
             decode_size: 2
             router_size: 1
-            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_1p1d_32p_aime2025.py
+            test_case: test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_1p1d_64p_aime2025.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'separation'
-          - name: test_npu_glm5_1_w4a8_1p1d_32p_in3k5_out1k5_50ms
+          - name: test_npu_glm5_1_w4a8_1p1d_64p_in3k5_out1k5_50ms
             prefill_size: 2
             decode_size: 2
             router_size: 1
-            test_case: test/registered/ascend/performance/test_npu_glm5_1_w4a8_1p1d_32p_in3k5_out1k5_50ms.py
+            test_case: test/registered/ascend/performance/test_npu_glm5_1_w4a8_1p1d_64p_in3k5_out1k5_50ms.py
             test_type: 'perf'
             prefill_decode_deployment: 'separation'
-          # Kimi 1p1d_16p (1P + 1D, each tp=16)
-          - name: test_npu_kimi_k2_5_w4a8_1p1d_16p_gpqa
+          # Kimi 1p1d_32p (P: tp=16 + D: tp=16, total 32 cards)
+          - name: test_npu_kimi_k2_5_w4a8_1p1d_32p_gpqa
             prefill_size: 1
             decode_size: 1
             router_size: 1
-            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_1p1d_16p_gpqa.py
+            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_1p1d_32p_gpqa.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'separation'
-          - name: test_npu_kimi_k2_5_w4a8_1p1d_16p_aime2025
+          - name: test_npu_kimi_k2_5_w4a8_1p1d_32p_aime2025
             prefill_size: 1
             decode_size: 1
             router_size: 1
-            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_1p1d_16p_aime2025.py
+            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_1p1d_32p_aime2025.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'separation'
-          - name: test_npu_kimi_k2_5_w4a8_1p1d_16p_in3k5_out1k5_50ms
+          - name: test_npu_kimi_k2_5_w4a8_1p1d_32p_in3k5_out1k5_50ms
             prefill_size: 1
             decode_size: 1
             router_size: 1
-            test_case: test/registered/ascend/performance/test_npu_kimi_k2_5_w4a8_1p1d_16p_in3k5_out1k5_50ms.py
+            test_case: test/registered/ascend/performance/test_npu_kimi_k2_5_w4a8_1p1d_32p_in3k5_out1k5_50ms.py
             test_type: 'perf'
             prefill_decode_deployment: 'separation'
-          # Kimi 2p1d_32p (2P(tp=16) + 1D(tp=32,nnodes=2))
-          - name: test_npu_kimi_k2_5_w4a8_2p1d_32p_gpqa
+          # Kimi 2p1d_64p (2P: tp=16 each + D: tp=32,nnodes=2, total 64 cards)
+          - name: test_npu_kimi_k2_5_w4a8_2p1d_64p_gpqa
             prefill_size: 2
             decode_size: 2
             router_size: 1
-            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_2p1d_32p_gpqa.py
+            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_2p1d_64p_gpqa.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'separation'
-          - name: test_npu_kimi_k2_5_w4a8_2p1d_32p_aime2025
+          - name: test_npu_kimi_k2_5_w4a8_2p1d_64p_aime2025
             prefill_size: 2
             decode_size: 2
             router_size: 1
-            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_2p1d_32p_aime2025.py
+            test_case: test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_2p1d_64p_aime2025.py
             test_type: 'accuracy'
             prefill_decode_deployment: 'separation'
-          - name: test_npu_kimi_k2_5_w4a8_2p1d_32p_in3k5_out1k5_50ms
+          - name: test_npu_kimi_k2_5_w4a8_2p1d_64p_in3k5_out1k5_50ms
             prefill_size: 2
             decode_size: 2
             router_size: 1
-            test_case: test/registered/ascend/performance/test_npu_kimi_k2_5_w4a8_2p1d_32p_in3k5_out1k5_50ms.py
+            test_case: test/registered/ascend/performance/test_npu_kimi_k2_5_w4a8_2p1d_64p_in3k5_out1k5_50ms.py
             test_type: 'perf'
             prefill_decode_deployment: 'separation'
     uses: ./.github/workflows/nightly-test-npu-e2e-multi-node.yml
