+++ test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_16p_aime2025.py
@@ -2,55 +2,53 @@ import unittest
 
 from sglang.test.ascend.e2e.test_npu_accuracy_utils import (
     BENCHMARK_TOOL_DEFAULT,
-    TestAscendAccuracyTestCaseBase,
+    TestAscendAccuracyMultiNodePdMixTestCaseBase,
 )
 from sglang.test.ascend.e2e.test_npu_performance_utils import GLM_5_1_W4A8_MODEL_PATH
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(
     est_time=3600,
-    suite="nightly-16-npu-a3",
+    suite="nightly-32-npu-a3",
     nightly=True,
 )
 
-GLM_5_1_SINGLE_NODE_ENVS = {
+GLM_5_1_TWO_NODE_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
     "STREAMS_PER_DEVICE": "32",
     "SGLANG_ENABLE_SPEC_V2": "1",
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "1",
-    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "64",
-    "HCCL_BUFFSIZE": "2000",
+    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "32",
+    "HCCL_BUFFSIZE": "2500",
 }
 
-GLM_5_1_SINGLE_NODE_OTHER_ARGS = [
+GLM_5_1_TWO_NODE_OTHER_ARGS = [
     "--attention-backend",
     "ascend",
     "--device",
     "npu",
     "--tp-size",
-    16,
+    32,
     "--nnodes",
-    1,
-    "--node-rank",
-    0,
+    2,
     "--dp-size",
     16,
     "--enable-dp-attention",
     "--chunked-prefill-size",
-    -1,
+    131072,
     "--max-prefill-tokens",
     280000,
     "--trust-remote-code",
     "--host",
     "0.0.0.0",
     "--mem-fraction-static",
-    0.8,
+    0.65,
     "--port",
     8001,
     "--served-model-name",
     "glm-5",
     "--cuda-graph-max-bs",
-    16,
+    8,
     "--max-running-requests",
     128,
     "--quantization",
@@ -73,22 +71,26 @@ GLM_5_1_SINGLE_NODE_OTHER_ARGS = [
     4,
 ]
 
+GLM_5_1_TWO_NODE_MODEL_CONFIG = {
+    "model_path": GLM_5_1_W4A8_MODEL_PATH,
+    "other_args": GLM_5_1_TWO_NODE_OTHER_ARGS,
+    "node_envs": GLM_5_1_TWO_NODE_ENVS,
+}
 
-class TestNPUGLM5_1_W4A8_16P_AIME2025(TestAscendAccuracyTestCaseBase):
-    """Test NPU accuracy for GLM-5.1-w4a8 16p single node on AIME 2025"""
 
+class TestNPUGLM5_1_W4A8_32P_AIME2025(TestAscendAccuracyMultiNodePdMixTestCaseBase):
+    """Test NPU accuracy for GLM-5.1-w4a8 32p two nodes on AIME 2025"""
+
+    model_config = GLM_5_1_TWO_NODE_MODEL_CONFIG
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
-    model = GLM_5_1_W4A8_MODEL_PATH
-    other_args = GLM_5_1_SINGLE_NODE_OTHER_ARGS
-    envs = GLM_5_1_SINGLE_NODE_ENVS
     accuracy = 0.8
     dataset_type = "aime2025"
     dataset_name = "aime2025_gen"
-    batch_size = 64
-    max_out_len = 8192
+    max_concurrency = 64
+    output_len = 8192
 
-    def test_npu_glm5_1_w4a8_16p_aime2025(self):
-        """Run NPU accuracy test for GLM-5.1-w4a8 single node on AIME 2025"""
+    def test_npu_glm5_1_w4a8_32p_aime2025(self):
+        """Run NPU accuracy test for GLM-5.1-w4a8 two nodes on AIME 2025"""
         self.run_accuracy()
