+++ test/registered/ascend/accuracy/test_npu_glm5_1_w4a8_8p_gpqa.py
@@ -2,53 +2,55 @@ import unittest
 
 from sglang.test.ascend.e2e.test_npu_accuracy_utils import (
     BENCHMARK_TOOL_DEFAULT,
-    TestAscendAccuracyMultiNodePdMixTestCaseBase,
+    TestAscendAccuracyTestCaseBase,
 )
 from sglang.test.ascend.e2e.test_npu_performance_utils import GLM_5_1_W4A8_MODEL_PATH
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(
     est_time=1800,
-    suite="nightly-32-npu-a3",
+    suite="nightly-16-npu-a3",
     nightly=True,
 )
 
-GLM_5_1_TWO_NODE_ENVS = {
+GLM_5_1_SINGLE_NODE_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
     "STREAMS_PER_DEVICE": "32",
     "SGLANG_ENABLE_SPEC_V2": "1",
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "1",
-    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "32",
-    "HCCL_BUFFSIZE": "2500",
+    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "64",
+    "HCCL_BUFFSIZE": "2000",
 }
 
-GLM_5_1_TWO_NODE_OTHER_ARGS = [
+GLM_5_1_SINGLE_NODE_OTHER_ARGS = [
     "--attention-backend",
     "ascend",
     "--device",
     "npu",
     "--tp-size",
-    32,
+    16,
     "--nnodes",
-    2,
+    1,
+    "--node-rank",
+    0,
     "--dp-size",
     16,
     "--enable-dp-attention",
     "--chunked-prefill-size",
-    131072,
+    -1,
     "--max-prefill-tokens",
     280000,
     "--trust-remote-code",
     "--host",
     "0.0.0.0",
     "--mem-fraction-static",
-    0.65,
+    0.8,
     "--port",
     8001,
     "--served-model-name",
     "glm-5",
     "--cuda-graph-max-bs",
-    8,
+    16,
     "--max-running-requests",
     128,
     "--quantization",
@@ -71,26 +73,22 @@ GLM_5_1_TWO_NODE_OTHER_ARGS = [
     4,
 ]
 
-GLM_5_1_TWO_NODE_MODEL_CONFIG = {
-    "model_path": GLM_5_1_W4A8_MODEL_PATH,
-    "other_args": GLM_5_1_TWO_NODE_OTHER_ARGS,
-    "node_envs": GLM_5_1_TWO_NODE_ENVS,
-}
 
+class TestNPUGLM5_1_W4A8_16P_GPQA(TestAscendAccuracyTestCaseBase):
+    """Test NPU accuracy for GLM-5.1-w4a8 16p single node on GPQA"""
 
-class TestNPUGLM5_1_W4A8_32P_GPQA(TestAscendAccuracyMultiNodePdMixTestCaseBase):
-    """Test NPU accuracy for GLM-5.1-w4a8 32p two nodes on GPQA"""
-
-    model_config = GLM_5_1_TWO_NODE_MODEL_CONFIG
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
+    model = GLM_5_1_W4A8_MODEL_PATH
+    other_args = GLM_5_1_SINGLE_NODE_OTHER_ARGS
+    envs = GLM_5_1_SINGLE_NODE_ENVS
     accuracy = 0.8
     dataset_type = "gpqa"
     dataset_name = "gpqa_gen_0_shot_cot_chat_prompt"
-    max_concurrency = 128
-    output_len = 1024
+    batch_size = 128
+    max_out_len = 1024
 
-    def test_npu_glm5_1_w4a8_32p_gpqa(self):
-        """Run NPU accuracy test for GLM-5.1-w4a8 two nodes on GPQA"""
+    def test_npu_glm5_1_w4a8_16p_gpqa(self):
+        """Run NPU accuracy test for GLM-5.1-w4a8 single node on GPQA"""
         self.run_accuracy()
