+++ test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_1p1d_32p_gpqa.py
@@ -127,8 +127,8 @@ MODEL_CONFIG = {
 }
 
 
-class TestNPUKimiK2_5_W4A8_1P1D_16P_GPQA(TestAscendAccuracyMultiNodePdSepTestCaseBase):
-    """Test NPU accuracy for Kimi-K2.5-w4a8 1p1d_16p on GPQA"""
+class TestNPUKimiK2_5_W4A8_1P1D_32P_GPQA(TestAscendAccuracyMultiNodePdSepTestCaseBase):
+    """Test NPU accuracy for Kimi-K2.5-w4a8 1p1d_32p on GPQA"""
 
     model_config = MODEL_CONFIG
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
@@ -138,8 +138,8 @@ class TestNPUKimiK2_5_W4A8_1P1D_16P_GPQA(TestAscendAccuracyMultiNodePdSepTestCas
     max_concurrency = 128
     output_len = 1024
 
-    def test_npu_kimi_k2_5_w4a8_1p1d_16p_gpqa(self):
-        """Run NPU accuracy test for Kimi-K2.5-w4a8 1p1d_16p on GPQA"""
+    def test_npu_kimi_k2_5_w4a8_1p1d_32p_gpqa(self):
+        """Run NPU accuracy test for Kimi-K2.5-w4a8 1p1d_32p on GPQA"""
         self.run_accuracy()
