+++ test/registered/ascend/accuracy/test_npu_kimi_k2_5_w4a8_2p1d_64p_gpqa.py
@@ -1,24 +1,23 @@
 import unittest
 
-from sglang.test.ascend.e2e.test_npu_performance_utils import (
-    AISBENCHMARK_DATASET_DEFAULT,
+from sglang.test.ascend.e2e.test_npu_accuracy_utils import (
     BENCHMARK_TOOL_DEFAULT,
-    KIMI_K2_5_W4A8_MODEL_PATH,
-    TestAscendPerfMultiNodePdSepTestCaseBase,
+    TestAscendAccuracyMultiNodePdSepTestCaseBase,
 )
+from sglang.test.ascend.e2e.test_npu_performance_utils import KIMI_K2_5_W4A8_MODEL_PATH
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(
     est_time=1800,
-    suite="nightly-pd-sep-2-node",
+    suite="nightly-pd-sep-4-node",
     nightly=True,
-    disabled="Currently it is executed by the npu performance workflow.",
 )
 
 PREFILL_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
     "STREAMS_PER_DEVICE": "32",
-    "HCCL_BUFFSIZE": "1800",
+    "HCCL_BUFFSIZE": "1600",
+    "SGLANG_DISAGGREGATION_BOOTSTRAP_TIMEOUT": "100",
 }
 
 DECODE_ENVS = {
@@ -26,9 +25,10 @@ DECODE_ENVS = {
     "STREAMS_PER_DEVICE": "32",
     "DEEP_NORMAL_MODE_USE_INT8_QUANT": "1",
     "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "32",
-    "HCCL_BUFFSIZE": "800",
+    "HCCL_BUFFSIZE": "2400",
     "SGLANG_ENABLE_SPEC_V2": "1",
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "1",
+    "SGLANG_DISAGGREGATION_BOOTSTRAP_TIMEOUT": "60",
 }
 
 PREFILL_ARGS = [
@@ -41,11 +41,11 @@ PREFILL_ARGS = [
     "--node-rank",
     0,
     "--mem-fraction-static",
-    0.75,
-    "--max-running-requests",
-    16,
+    0.76,
     "--chunked-prefill-size",
-    32768,
+    16384,
+    "--context-length",
+    133120,
     "--quantization",
     "modelslim",
     "--disaggregation-transfer-backend",
@@ -74,13 +74,18 @@ DECODE_ARGS = [
     "--disaggregation-mode",
     "decode",
     "--tp-size",
-    16,
+    32,
     "--nnodes",
-    1,
+    2,
     "--mem-fraction-static",
-    0.76,
+    0.67,
     "--max-running-requests",
-    16,
+    32,
+    "--chunked-prefill-size",
+    65536,
+    "--context-length",
+    133120,
+    "--disable-radix-cache",
     "--quantization",
     "modelslim",
     "--disaggregation-transfer-backend",
@@ -99,11 +104,10 @@ DECODE_ARGS = [
     "--deepep-mode",
     "auto",
     "--dp-size",
-    4,
+    32,
     "--enable-dp-attention",
     "--cuda-graph-bs",
-    4,
-    8,
+    1,
     "--dtype",
     "bfloat16",
     "--speculative-draft-model-quantization",
@@ -124,29 +128,25 @@ MODEL_CONFIG = {
     "decode_args": DECODE_ARGS,
     "prefill_envs": PREFILL_ENVS,
     "decode_envs": DECODE_ENVS,
-    "router_args": ["--policy", "round_robin"],
+    "router_args": ["--policy", "cache_aware"],
     "router_envs": {},
 }
 
 
-class TestNPUKimiK2_5_W4A8_1P1D_16P_In3k5_Out1k5_50ms(TestAscendPerfMultiNodePdSepTestCaseBase):
-    """Test NPU performance for Kimi-K2.5-w4a8 1p1d_16p PD separation"""
+class TestNPUKimiK2_5_W4A8_2P1D_64P_GPQA(TestAscendAccuracyMultiNodePdSepTestCaseBase):
+    """Test NPU accuracy for Kimi-K2.5-w4a8 2p1d_64p on GPQA"""
 
     model_config = MODEL_CONFIG
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
-    aisbench_dataset_type = AISBENCHMARK_DATASET_DEFAULT
-    dataset_name = "random"
+    accuracy = 0.8
+    dataset_type = "gpqa"
+    dataset_name = "gpqa_gen_0_shot_cot_chat_prompt"
     max_concurrency = 128
-    num_prompts = 512
-    input_len = 3500
-    output_len = 1500
-    random_range_ratio = 1
-    tpot = 50
-    output_token_throughput = 3000
+    output_len = 1024
 
-    def test_npu_kimi_k2_5_w4a8_1p1d_16p_in3k5_out1k5_50ms(self):
-        """Run NPU performance test for Kimi-K2.5-w4a8 1p1d_16p"""
-        self.run_throughput()
+    def test_npu_kimi_k2_5_w4a8_2p1d_64p_gpqa(self):
+        """Run NPU accuracy test for Kimi-K2.5-w4a8 2p1d_64p on GPQA"""
+        self.run_accuracy()
 
 
 if __name__ == "__main__":
