+++ test/registered/ascend/performance/test_npu_glm5_1_w4a8_1p1d_64p_in3k5_out1k5_50ms.py
@@ -1,16 +1,18 @@
 import unittest
 
-from sglang.test.ascend.e2e.test_npu_accuracy_utils import (
+from sglang.test.ascend.e2e.test_npu_performance_utils import (
+    AISBENCHMARK_DATASET_DEFAULT,
     BENCHMARK_TOOL_DEFAULT,
-    TestAscendAccuracyMultiNodePdSepTestCaseBase,
+    GLM_5_1_W4A8_MODEL_PATH,
+    TestAscendPerfMultiNodePdSepTestCaseBase,
 )
-from sglang.test.ascend.e2e.test_npu_performance_utils import GLM_5_1_W4A8_MODEL_PATH
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(
-    est_time=3600,
+    est_time=1800,
     suite="nightly-pd-sep-4-node",
     nightly=True,
+    disabled="Currently it is executed by the npu performance workflow.",
 )
 
 GLM_5_1_PD_SEP_PREFILL_ENVS = {
@@ -163,20 +165,24 @@ GLM_5_1_PD_SEP_MODEL_CONFIG = {
 }
 
 
-class TestNPUGLM5_1_W4A8_PD_SEP_AIME2025(TestAscendAccuracyMultiNodePdSepTestCaseBase):
-    """Test NPU accuracy for GLM-5.1-w4a8 PD separation 4 nodes on AIME 2025"""
+class TestNPUGLM5_1_W4A8_1P1D_64P_In3k5_Out1k5_50ms(TestAscendPerfMultiNodePdSepTestCaseBase):
+    """Test NPU performance for GLM-5.1-w4a8 1p1d_64p PD separation in3k5 out1k5"""
 
     model_config = GLM_5_1_PD_SEP_MODEL_CONFIG
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
-    accuracy = 0.8
-    dataset_type = "aime2025"
-    dataset_name = "aime2025_gen"
-    max_concurrency = 64
-    output_len = 8192
+    aisbench_dataset_type = AISBENCHMARK_DATASET_DEFAULT
+    dataset_name = "random"
+    max_concurrency = 128
+    num_prompts = 512
+    input_len = 3500
+    output_len = 1500
+    random_range_ratio = 1
+    tpot = 50
+    output_token_throughput = 3000
 
-    def test_npu_glm5_1_w4a8_pd_sep_aime2025(self):
-        """Run NPU accuracy test for GLM-5.1-w4a8 PD separation on AIME 2025"""
-        self.run_accuracy()
+    def test_npu_glm5_1_w4a8_1p1d_64p_in3k5_out1k5_50ms(self):
+        """Run NPU performance test for GLM-5.1-w4a8 1p1d_64p"""
+        self.run_throughput()
 
 
 if __name__ == "__main__":
