+++ test/registered/ascend/performance/test_npu_kimi_k2_5_w4a8_16p_in3k5_out1k5_50ms.py
@@ -23,12 +23,10 @@ KIMI_K2_5_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
     "STREAMS_PER_DEVICE": "32",
     "DEEP_NORMAL_MODE_USE_INT8_QUANT": "1",
-    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "96",
+    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "64",
     "HCCL_BUFFSIZE": "1200",
     "SGLANG_ENABLE_SPEC_V2": "1",
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "1",
-    "SGLANG_SCHEDULER_DECREASE_PREFILL_IDLE": "1",
-    "SGLANG_PREFILL_DELAYER_MAX_DELAY_PASSES": "200",
 }
 
 KIMI_K2_5_OTHER_ARGS = [
@@ -53,11 +51,11 @@ KIMI_K2_5_OTHER_ARGS = [
     "--base-gpu-id",
     0,
     "--mem-fraction-static",
-    0.78,
+    0.8,
     "--max-running-requests",
-    160,
+    128,
     "--chunked-prefill-size",
-    32768,
+    16384,
     "--context-length",
     8192,
     "--max-prefill-tokens",
@@ -73,25 +71,22 @@ KIMI_K2_5_OTHER_ARGS = [
     "auto",
     "--enable-dp-attention",
     "--dp-size",
-    16,
+    4,
     "--cuda-graph-bs",
-    1,
-    2,
     4,
     8,
-    9,
-    10,
-    "--disable-radix-cache",
+    16,
+    32,
     "--speculative-algorithm",
     "EAGLE3",
     "--speculative-draft-model-path",
     KIMI_K2_5_EAGLE3_MODEL_PATH,
     "--speculative-num-steps",
-    4,
+    3,
     "--speculative-eagle-topk",
     1,
     "--speculative-num-draft-tokens",
-    5,
+    4,
     "--speculative-draft-model-quantization",
     "unquant",
 ]
