+++ test/registered/ascend/performance/test_npu_kimi_k2_5_w4a8_1p1d_32p_in3k5_out1k5_50ms.py
@@ -1,23 +1,24 @@
 import unittest
 
-from sglang.test.ascend.e2e.test_npu_accuracy_utils import (
+from sglang.test.ascend.e2e.test_npu_performance_utils import (
+    AISBENCHMARK_DATASET_DEFAULT,
     BENCHMARK_TOOL_DEFAULT,
-    TestAscendAccuracyMultiNodePdSepTestCaseBase,
+    KIMI_K2_5_W4A8_MODEL_PATH,
+    TestAscendPerfMultiNodePdSepTestCaseBase,
 )
-from sglang.test.ascend.e2e.test_npu_performance_utils import KIMI_K2_5_W4A8_MODEL_PATH
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(
-    est_time=3600,
-    suite="nightly-pd-sep-4-node",
+    est_time=1800,
+    suite="nightly-pd-sep-2-node",
     nightly=True,
+    disabled="Currently it is executed by the npu performance workflow.",
 )
 
 PREFILL_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
     "STREAMS_PER_DEVICE": "32",
-    "HCCL_BUFFSIZE": "1600",
-    "SGLANG_DISAGGREGATION_BOOTSTRAP_TIMEOUT": "100",
+    "HCCL_BUFFSIZE": "1800",
 }
 
 DECODE_ENVS = {
@@ -25,10 +26,9 @@ DECODE_ENVS = {
     "STREAMS_PER_DEVICE": "32",
     "DEEP_NORMAL_MODE_USE_INT8_QUANT": "1",
     "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "32",
-    "HCCL_BUFFSIZE": "2400",
+    "HCCL_BUFFSIZE": "800",
     "SGLANG_ENABLE_SPEC_V2": "1",
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "1",
-    "SGLANG_DISAGGREGATION_BOOTSTRAP_TIMEOUT": "60",
 }
 
 PREFILL_ARGS = [
@@ -41,11 +41,11 @@ PREFILL_ARGS = [
     "--node-rank",
     0,
     "--mem-fraction-static",
-    0.76,
+    0.75,
+    "--max-running-requests",
+    16,
     "--chunked-prefill-size",
-    16384,
-    "--context-length",
-    133120,
+    32768,
     "--quantization",
     "modelslim",
     "--disaggregation-transfer-backend",
@@ -74,18 +74,13 @@ DECODE_ARGS = [
     "--disaggregation-mode",
     "decode",
     "--tp-size",
-    32,
+    16,
     "--nnodes",
-    2,
+    1,
     "--mem-fraction-static",
-    0.67,
+    0.76,
     "--max-running-requests",
-    32,
-    "--chunked-prefill-size",
-    65536,
-    "--context-length",
-    133120,
-    "--disable-radix-cache",
+    16,
     "--quantization",
     "modelslim",
     "--disaggregation-transfer-backend",
@@ -104,10 +99,11 @@ DECODE_ARGS = [
     "--deepep-mode",
     "auto",
     "--dp-size",
-    32,
+    4,
     "--enable-dp-attention",
     "--cuda-graph-bs",
-    1,
+    4,
+    8,
     "--dtype",
     "bfloat16",
     "--speculative-draft-model-quantization",
@@ -128,25 +124,29 @@ MODEL_CONFIG = {
     "decode_args": DECODE_ARGS,
     "prefill_envs": PREFILL_ENVS,
     "decode_envs": DECODE_ENVS,
-    "router_args": ["--policy", "cache_aware"],
+    "router_args": ["--policy", "round_robin"],
     "router_envs": {},
 }
 
 
-class TestNPUKimiK2_5_W4A8_2P1D_32P_AIME2025(TestAscendAccuracyMultiNodePdSepTestCaseBase):
-    """Test NPU accuracy for Kimi-K2.5-w4a8 2p1d_32p on AIME 2025"""
+class TestNPUKimiK2_5_W4A8_1P1D_32P_In3k5_Out1k5_50ms(TestAscendPerfMultiNodePdSepTestCaseBase):
+    """Test NPU performance for Kimi-K2.5-w4a8 1p1d_32p PD separation in3k5 out1k5"""
 
     model_config = MODEL_CONFIG
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
-    accuracy = 0.8
-    dataset_type = "aime2025"
-    dataset_name = "aime2025_gen"
-    max_concurrency = 64
-    output_len = 8192
+    aisbench_dataset_type = AISBENCHMARK_DATASET_DEFAULT
+    dataset_name = "random"
+    max_concurrency = 128
+    num_prompts = 512
+    input_len = 3500
+    output_len = 1500
+    random_range_ratio = 1
+    tpot = 50
+    output_token_throughput = 3000
 
-    def test_npu_kimi_k2_5_w4a8_2p1d_32p_aime2025(self):
-        """Run NPU accuracy test for Kimi-K2.5-w4a8 2p1d_32p on AIME 2025"""
-        self.run_accuracy()
+    def test_npu_kimi_k2_5_w4a8_1p1d_32p_in3k5_out1k5_50ms(self):
+        """Run NPU performance test for Kimi-K2.5-w4a8 1p1d_32p"""
+        self.run_throughput()
 
 
 if __name__ == "__main__":
